+++ sql_inits/create_all.sql
@@ -517,7 +517,7 @@ ORDER BY (ts) TTL toDateTime(ts) + toIntervalDay(30);
 
 CREATE TABLE SystemService.PRF(
   db_uid LowCardinality(String),
-  operation_target_value Int16, 
+  operation_target_value Float,
   operation_priority Int16,
   file LowCardinality(String),
   conf_name LowCardinality(String),
@@ -531,6 +531,4 @@ CREATE TABLE SystemService.PRF(
   user LowCardinality(String),
 ) ENGINE = MergeTree() PARTITION BY toYYYYMM(ts)
 ORDER BY (ts) TTL toDateTime(ts) + toIntervalDay(30);
-
-
